+++ tests/test_app/migrations/0002_fakeproductnoprice.py
@@ -1,0 +1,21 @@
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ("test_app", "0001_initial"),
+    ]
+
+    operations = [
+        migrations.CreateModel(
+            name="FakeProductNoPrice",
+            fields=[
+                ("id", models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name="ID")),
+                ("name", models.CharField(max_length=100)),
+            ],
+            options={
+                "app_label": "test_app",
+            },
+        ),
+    ]